reorganized the display functions into a new module
