+++ utils/plot_handler.py
@@ -10,13 +10,13 @@ Classes:
 Functions:
     plot_band_structure: Plot band structure from processed data.
     plot_wannier_comparison: Plot Wannier and DFT band structure comparison.
+    plot_pdos: Plot projected density of states (PDOS) data.
 """
 import os
 from sys import argv
 from typing import Dict, List
 import matplotlib.pyplot as plt
 import numpy as np
-from rich import box
 from rich.table import Table
 
 from core.config import BandsPlotConfig, DOSPlotConfig
@@ -25,6 +25,7 @@ from data.data_collector import prepare_bands_info, prepare_wannier_info, prepar
 from data.models import ProjectSetup
 from core.project_setup import initialize_project
 from plotting.plotters import BandPlotter, WannierComparePlotter, DOSPlotter
+from ui.plot_info_display import display_band_plot_info, display_wannier_plot_info, display_pdos_plot_info
 from ui.ui_helpers import print_header, console, print_success, prompt_input
 from utils.print_thanks import print_animated_ascii
 
@@ -208,86 +209,6 @@ class ProjectionDataProcessor:
                 combined_weights[projection] = weight
 
         return combined_weights
-
-
-def display_band_plot_info(projection_info: Dict,
-                           spin_orbit_flag: str,
-                           include_stress: bool = False,
-                           stress_amount: str = None):
-    """
-    Display debug information for band structure plots in a formatted table.
-
-    Args:
-        projection_info (Dict): Dictionary containing projection data for each element.
-        spin_orbit_flag (str): Flag indicating whether spin-orbit coupling (SOC) is enabled.
-        include_stress (bool, optional): Whether to include stress information in the table. Defaults to False.
-        stress_amount (str, optional): Stress amount as a string (e.g., '1_30'). Defaults to None.
-    """
-    # Creating a table with a title and rounded box style
-    debug_table = Table(title="Projection Info Debug", box=box.ROUNDED)
-    debug_table.add_column("Parameter", style="cyan")
-    debug_table.add_column("Value", style="green")
-
-    # Add strain information if stress is included, otherwise add SOC status
-    if include_stress:
-        strain = float(stress_amount.replace('_', '.')) * 100
-        debug_table.add_row("Strain", f"{strain:.2f}%")
-    else:
-        debug_table.add_row("SOC", "Yes" if spin_orbit_flag else "No")
-
-    # Add rows for each element's projection data, excluding orbital weights
-    for atom, info in projection_info.items():
-        debug_info = {k: v for k, v in info.items() if k != "orbital_weights"}
-        debug_table.add_row(f"Element {atom}", str(debug_info))
-
-    console.print(debug_table)
-
-
-def display_wannier_plot_info(fermi_energy: float, alat: float, flag: str):
-    """
-    Display debug information for Wannier comparison plots in a formatted table.
-
-    Args:
-        fermi_energy (float): Fermi energy value in eV.
-        alat (float): Lattice parameter in Å.
-        flag (str): Spin-orbit coupling (SOC) flag (e.g., '_soc').
-    """
-    # Creating a table with a title and rounded box style
-    debug_table = Table(box=box.ROUNDED, title="Wannier Comparison Debug Info")
-    debug_table.add_column("Parameter", style="cyan")
-    debug_table.add_column("Value", style="green")
-
-    debug_table.add_row("Fermi Energy (eV)", f"{fermi_energy:.4f}")
-    debug_table.add_row("Alat (Å)", f"{alat:.6f}")
-    debug_table.add_row("SOC Flag", "Yes" if flag == "_soc" else "No")
-
-    console.print(debug_table)
-
-
-def display_pdos_plot_info(elements: List[str], projection_data: Dict[str, Dict]):
-    """
-    Display debug information for projected density of states (PDOS) plots in a formatted table.
-
-    Args:
-        elements (List[str]): List of unique chemical elements.
-        projection_data (Dict[str, Dict]): Dictionary containing projection data for each element.
-    """
-    # Creating a table with a title and rounded box style
-    debug_table = Table(box=box.ROUNDED, title="PDOS Debug Info")
-    debug_table.add_column("Parameter", style="cyan")
-    debug_table.add_column("Value", style="green")
-
-    debug_table.add_row("Elements", ", ".join(elements))
-
-    for element, data in projection_data.items():
-        element_info = {
-            "index": data["index"],
-            "projected_orbitals": data["projected_orbitals"],
-            "plot_colors": data["plot_colors"]
-        }
-        debug_table.add_row(f"Element {element}", str(element_info))
-
-    console.print(debug_table)
 
 
 def plot_band_structure(project: ProjectSetup,
@@ -612,7 +533,7 @@ if __name__ == "__main__":
             prepare_wannier_info(project)
             process_comparison_data(project)
 
-            plot_config = BandsPlotConfig()
+            plot_config = BandsPlotConfig.get_default_config()
             plot_wannier_comparison(project, plot_config, save_fig=False, test_module=True)
 
         case "bands":
@@ -620,7 +541,7 @@ if __name__ == "__main__":
             prepare_bands_info(project)
             process_band_data(project)
 
-            plot_config = BandsPlotConfig()
+            plot_config = BandsPlotConfig.get_default_config()
             plot_band_structure(project, plot_config, save_fig=False, test_module=True)
 
         case "pdos":
@@ -628,7 +549,7 @@ if __name__ == "__main__":
             prepare_pdos_info(project)
             process_pdos_data(project)
 
-            plot_config = DOSPlotConfig()
+            plot_config = DOSPlotConfig.get_default_config()
             plot_pdos(project, plot_config, save_fig=False, test_module=False)
 
         case _:
